mysite/requirements.txt
asgiref==3.2.3
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
chardet2==2.0.3
Django==3.0
django-cors-headers==3.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
httplib2==0.14.0
oauth2==1.9.0.post1
oauthlib==3.1.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyJWT==1.7.1
pytz==2019.3
requests==0.13.1
requests-oauthlib==1.3.0
rsa==4.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
vimeo==0.1.9
